// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  VIEWER: 'VIEWER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const PriceField = {
  UAE_PRICE: 'UAE_PRICE',
  IR_PRICE: 'IR_PRICE'
} as const

export type PriceField = (typeof PriceField)[keyof typeof PriceField]


export const SyncStatus = {
  PENDING: 'PENDING',
  SUCCEEDED: 'SUCCEEDED',
  FAILED: 'FAILED'
} as const

export type SyncStatus = (typeof SyncStatus)[keyof typeof SyncStatus]